requirements_personaplex.txt
# LiveKit
livekit==1.0.23
livekit-agents==1.3.12
livekit-api==1.1.0
livekit-plugins-silero==1.3.12   # VAD only

# WebSocket
websockets==15.0.1

# Audio math
numpy==2.2.6

# Config
python-dotenv==1.2.1

# Logging
coloredlogs==15.0.1
fastapi
uvicorn

scipy>=1.11.0
sphn==0.1.4

# ── Transcript (NEW) ───────────────────────────────────────────────────────
# Local Whisper — no API key, runs on CPU or GPU
# CPU: ~0.5-1s per utterance with 'base' model
# GPU: real-time with 'small' model
faster-whisper>=1.0.0

requests

pydantic>=2.0.0
sendgrid>=4.0.0
deepgram-sdk==3.10.1

httpx>=0.27.0